split util into separate submodules
