+++ openzwavemqtt/util/__init__.py
@@ -1,0 +1,1 @@
+"""Utility module for OpenZWave."""
