+++ openzwavemqtt/util/node.py
@@ -1,13 +1,10 @@
-"""Utility functions and classes for OpenZWave."""
+"""Utility functions for OpenZWave nodes."""
 from typing import cast, Dict, List, Union
 
-from .const import (
-    ATTR_CODE_SLOT,
-    ATTR_IN_USE,
+from ..const import (
     ATTR_LABEL,
     ATTR_MAX,
     ATTR_MIN,
-    ATTR_NAME,
     ATTR_OPTIONS,
     ATTR_PARAMETER,
     ATTR_POSITION,
@@ -15,13 +12,12 @@ from .const import (
     ATTR_VALUE,
     CommandClass,
     ValueGenre,
-    ValueIndex,
     ValueType,
 )
-from .exceptions import InvalidValueError, NotFoundError, WrongTypeError
-from .manager import OZWManager
-from .models.node import OZWNode
-from .models.value import OZWValue
+from ..exceptions import InvalidValueError, NotFoundError, WrongTypeError
+from ..manager import OZWManager
+from ..models.node import OZWNode
+from ..models.value import OZWValue
 
 
 def get_node_from_manager(
@@ -236,46 +232,3 @@ def get_config_parameters(
         values.append(value_to_return)
 
     return values
-
-
-def get_code_slots(node: OZWNode) -> List[Dict[str, Union[int, bool, str]]]:
-    """Get all code slots on the lock and whether or not they are used."""
-    command_class = node.get_command_class(CommandClass.USER_CODE)
-
-    if not command_class:
-        raise NotFoundError("Entity doesn't have any user codes available to configure")
-
-    return [
-        {
-            ATTR_CODE_SLOT: value.index.value,
-            ATTR_NAME: value.label,
-            ATTR_IN_USE: value.value_set,
-        }
-        for value in command_class.values()  # type: ignore
-        if value.genre == ValueGenre.USER
-    ]
-
-
-def set_usercode(node: OZWNode, code_slot: int, usercode: str) -> None:
-    """Set the usercode to index X on the lock."""
-    value = node.get_value(CommandClass.USER_CODE, code_slot)
-
-    if not value:
-        raise NotFoundError(f"Code slot {code_slot} not found")
-
-    if len(str(usercode)) < 4:
-        raise InvalidValueError("User code must be at least 4 digits")
-
-    value.send_value(usercode)  # type: ignore
-
-
-def clear_usercode(node: OZWNode, code_slot: int) -> None:
-    """Clear usercode in slot X on the lock."""
-    value = node.get_value(CommandClass.USER_CODE, ValueIndex.CLEAR_USER_CODE)
-
-    if not value:
-        raise WrongTypeError("Entity is not capable of clearing user codes")
-
-    value.send_value(code_slot)  # type: ignore
-    # Sending twice because the first time it doesn't take
-    value.send_value(code_slot)  # type: ignore
